added new migration `changed coolerHeight from int to double`
+ fixed some mistakes in the interfaces
+ added input validations in new viewModel for `saveConfigs`
+ added idea for calculating `configPrice`

+++ PCHCB/PCHCB.Data.Models/Case.cs
@@ -54,7 +54,7 @@
         /// Case Max Air CPU Cooler Height (in mm)
         /// </summary>
         [Required]
-        public int MaxAirCpuCoolerHeight { get; set; }
+        public double MaxAirCpuCoolerHeight { get; set; }
 
         /// <summary>
         /// PSU Factor (ATX, SFX, etc.)
